be/pip_requirements.txt
django-currentuser==0.2.3
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.5
